backend/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.8
colorama==0.4.6
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.3
sympy==1.12
tomli==2.2.1
typing-extensions==4.12.2
tzdata==2024.2
